chore: scale folder structure to support multichain for tests and contracts
