+++ .env.example
@@ -1,5 +1,4 @@
 RPC_URL=https://mainnet.infura.io/v3/yourInfuraId
 RPC_URL_ARBITRUM=https://arbitrum-mainnet.infura.io/v3/yourInfuraId
-TEST_CHAIN_FORK=mainnet
 PRIVATE_KEY=0xPrivateKey
 ETHERSCAN_API_KEY=yourEtherscanApiKey
