+++ contracts/mainnet/SaddleThreeTokens.sol
@@ -1,9 +1,9 @@
 pragma solidity ^0.8.6;
 
-import "./interfaces/ISaddlePool.sol";
-import "./interfaces/IERC20.sol";
-import "./interfaces/ITrigger.sol";
-import "./interfaces/IYVaultV2.sol";
+import "../shared/interfaces/ISaddlePool.sol";
+import "../shared/interfaces/IERC20.sol";
+import "../shared/interfaces/ITrigger.sol";
+import "../shared/interfaces/IYVaultV2.sol";
 
 /**
  * @notice Defines a trigger that is toggled if any of the following conditions occur:
