+++ contracts/shared/interfaces/ICurvePool.sol
