+++ contracts/shared/interfaces/IRibbonVaultV2.sol
