+++ contracts/shared/interfaces/ITrigger.sol
