+++ contracts/shared/interfaces/IYVaultV2.sol
