+++ package.json
@@ -58,7 +58,9 @@
     "lint:sol": "solhint --config ./.solhint.json --max-warnings 0 \"contracts/**/*.sol\"",
     "lint:ts": "eslint --config ./.eslintrc.yaml --ignore-path ./.eslintignore --ext .js,.ts .",
     "prettier": "prettier --config .prettierrc --write \"**/*.{js,json,md,sol,js,ts}\"",
-    "test": "yarn compile && hardhat test",
+    "test": "yarn test:mainnet; yarn test:arbitrum",
+    "test:mainnet": "yarn compile && TEST_CHAIN_FORK=mainnet hardhat test test/mainnet/**",
+    "test:arbitrum": "yarn compile && TEST_CHAIN_FORK=arbitrum hardhat test test/arbitrum/**",
     "typechain": "hardhat typechain",
     "localhost": "source .env && yarn hardhat node --fork $RPC_URL"
   },
