+++ test/mainnet/RibbonV2SharePrice.test.ts
@@ -1,63 +1,74 @@
 // --- Imports ---
 import { artifacts, ethers, network, waffle } from 'hardhat';
 import { expect } from 'chai';
-import type { BigNumber, BigNumberish } from '@ethersproject/bignumber';
+import type { BigNumberish } from '@ethersproject/bignumber';
+import { keccak256 } from '@ethersproject/keccak256';
 import { SignerWithAddress } from '@nomiclabs/hardhat-ethers/dist/src/signer-with-address';
-import { MockCozyToken, IERC20, IRariVault, RariSharePrice } from '../typechain';
+import { MockCozyToken, IRibbonVaultV2, RibbonV2SharePrice, IERC20 } from '../../typechain';
 
 // --- Constants and extracted methods ---
-const wad = 10n ** 18n;
 const { deployContract, loadFixture } = waffle;
 const { MaxUint256: MAX_UINT } = ethers.constants;
-const { hexZeroPad } = ethers.utils;
+const { defaultAbiCoder, hexZeroPad, hexStripZeros } = ethers.utils;
 const BN = (x: BigNumberish) => ethers.BigNumber.from(x);
 const to32ByteHex = (x: BigNumberish) => hexZeroPad(BN(x).toHexString(), 32);
 
 interface VaultInfo {
   name: string;
-  market: string;
+  ribbonVault: string;
 }
 
-// Rari vaults
-const markets: VaultInfo[] = [
-  { name: 'Rari USDC Pool', market: '0xC6BF8C8A55f77686720E0a88e2Fd1fEEF58ddf4a' },
-  { name: 'Rari DAI Pool', market: '0xB465BAF04C087Ce3ed1C266F96CA43f4847D9635' },
+// Ribbon vault pairings to test
+const vaults: VaultInfo[] = [
+  {
+    name: 'T-ETH-C',
+    ribbonVault: '0x25751853Eab4D0eB3652B5eB6ecB102A2789644B',
+  },
 ];
 
+// --- Storage slot helper methods ---
+// `defaultAbiCoder.encode` is equivalent to Solidity's `abi.encode()`, and we strip leading zeros from the hashed
+// value to conform to the JSON-RPC spec: https://ethereum.org/en/developers/docs/apis/json-rpc/#hex-value-encoding
+
+// Returns the storage slot for a Solidity mapping from an `address` to a value, given the slot of the mapping itself,
+//  `mappingSlot`. Read more at https://docs.soliditylang.org/en/latest/internals/layout_in_storage.html#mappings-and-dynamic-arrays
+const getSolidityStorageSlot = (mappingSlot: string, address: string) => {
+  return hexStripZeros(keccak256(defaultAbiCoder.encode(['address', 'uint256'], [address, mappingSlot])));
+};
+
+// Returns the storage slot for a Vyper mapping from an `address` to a value, given the slot of the mapping itself,
+//  `mappingSlot`. May be dependent on the Vyper version used. See this tweet thread for more info: https://twitter.com/msolomon44/status/1420137730009300992
+const getVyperStorageSlot = (mappingSlot: string, address: string) => {
+  return hexStripZeros(keccak256(defaultAbiCoder.encode(['uint256', 'address'], [mappingSlot, address])));
+};
+
 // --- Tests ---
-markets.forEach((vault) => {
-  describe(`Vault: ${vault.name}`, function () {
+vaults.forEach((vault) => {
+  describe(`Ribbon Vault: ${vault.name}`, function () {
     // --- Data ---
-    let market: IRariVault;
-    let token: IERC20;
+    let ribbonVault: IRibbonVaultV2;
     let deployer: SignerWithAddress;
-    let trigger: RariSharePrice;
+    let trigger: RibbonV2SharePrice;
     let triggerParams: (string | number[])[] = []; // trigger params deployment parameters
 
     // --- Functions modifying storage ---
     /**
-     * @notice Change Rari vault's price per share by modifying total supply (since share price is a getter method that
+     * @notice Change Ribbon vault's price per share by modifying total supply (since share price is a getter method that
      * divides by total token supply)
      * @param supply New total supply. To zero out share price, use MAX_UINT256, which simulates unlimited minting of
      * shares, making price per share effectively 0
      */
-    async function setVaultTotalSupply(supply: BigNumberish) {
-      const storageSlot = '0x35';
-      await network.provider.send('hardhat_setStorageAt', [token.address, storageSlot, to32ByteHex(supply)]);
+    async function setRibbonTotalSupply(supply: BigNumberish) {
+      const storageSlot = '0x99'; // storage slot 153 in Ribbon vault contains total supply (lots of gap due to OpenZeppelin's implementation)
+      await network.provider.send('hardhat_setStorageAt', [vault.ribbonVault, storageSlot, to32ByteHex(supply)]);
     }
 
     // --- Test fixture ---
+
     // Executes checkAndToggleTrigger and verifies the expected state
     async function assertTriggerStatus(status: boolean) {
       await trigger.checkAndToggleTrigger();
       expect(await trigger.isTriggered()).to.equal(status);
-    }
-
-    // Gets the effective share price of the vault
-    async function getPricePerShare() {
-      const balance = <BigNumber>await market.callStatic.getFundBalance();
-      const supply = <BigNumber>await token.totalSupply();
-      return balance.mul(wad).div(supply); // need to scale balance before division since both balance and supply are on the scale of 1e18
     }
 
     /**
@@ -70,93 +81,27 @@ markets.forEach((vault) => {
       const [deployer, recipient] = await ethers.getSigners();
 
       // Get mainnet contract instances
-      const market = <IRariVault>await ethers.getContractAt('IRariVault', vault.market);
-      const token = <IERC20>await ethers.getContractAt('IERC20', await market.rariFundToken());
+      const ribbonVault = <IRibbonVaultV2>await ethers.getContractAt('IRibbonVaultV2', vault.ribbonVault);
 
-      // Deploy trigger
+      // Deploy RibbonV2SharePrice trigger
       const triggerParams = [
-        'Rari USDC Trigger', // name
-        'rariUSDC-TRIG', // symbol
-        'Triggers when the Rari USDC vault share price decreases by more than 50% between consecutive checks.', // description
-        [10], // platform ID
+        'Ribbon T-ETH-C Trigger', // name
+        'RibbonT-ETH-C-TRIG', // symbol
+        'Triggers when the Ribbon Theta Vault ETH pool virtual price decreases by more than 50% between consecutive checks', // description
+        [9], // platform ID for ribbon
         recipient.address, // subsidy recipient
-        vault.market, // mainnet address
+        vault.ribbonVault, // mainnet ribbon vault
       ];
 
-      const triggerArtifact = await artifacts.readArtifact('RariSharePrice');
-      const trigger = <RariSharePrice>await deployContract(deployer, triggerArtifact, triggerParams);
+      const RibbonV2SharePriceArtifact = await artifacts.readArtifact('RibbonV2SharePrice');
+      const trigger = <RibbonV2SharePrice>await deployContract(deployer, RibbonV2SharePriceArtifact, triggerParams);
 
-      return { deployer, market, token, trigger, triggerParams };
+      return { deployer, ribbonVault, trigger, triggerParams };
     }
 
     // --- Tests ---
     beforeEach(async () => {
-      ({ deployer, market, token, trigger, triggerParams } = await loadFixture(setupFixture));
-    });
-
-    it('USDC pool: share price only increases', async () => {
-      // --- Helper variables and methods ---
-      if (vault.name !== 'Rari USDC Pool') return;
-      const usdcAddress = '0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48';
-      const usdc = new ethers.Contract(usdcAddress, ['function approve(address,uint256) external'], deployer);
-
-      const rariAbi = [
-        'function deposit(string calldata currencyCode, uint256 amount) external',
-        'function withdraw(string calldata currencyCode, uint256 amount) external returns (uint256)',
-      ];
-      const rari = new ethers.Contract(market.address, rariAbi, deployer);
-      const rariToken = await ethers.getContractAt('IERC20', '0x016bf078ABcaCB987f0589a6d3BEAdD4316922B0', deployer);
-
-      function getBalanceOfSlotSolidity(mappingSlot: string, address: string) {
-        const { hexStripZeros, keccak256, defaultAbiCoder } = ethers.utils;
-        return hexStripZeros(keccak256(defaultAbiCoder.encode(['address', 'uint256'], [address, mappingSlot])));
-      }
-
-      async function setUsdcBalance(account: string, amount: BigNumberish) {
-        const slot = getBalanceOfSlotSolidity('0x9', account);
-        await network.provider.send('hardhat_setStorageAt', [usdcAddress, slot, to32ByteHex(amount)]);
-      }
-
-      async function getPricePerShare() {
-        await trigger.checkAndToggleTrigger();
-        return trigger.lastPricePerShare();
-      }
-
-      async function getRariBalance() {
-        return rariToken.balanceOf(deployer.address);
-      }
-
-      // --- Setup ---
-      const sp1 = await getPricePerShare();
-      const bal1 = await getRariBalance();
-
-      // --- Give user USDC ---
-      const amount = ethers.utils.parseUnits('1000000000', 6);
-      await setUsdcBalance(deployer.address, amount);
-
-      // --- Deposit ---
-      await usdc.approve(rari.address, ethers.constants.MaxUint256);
-      await rari.deposit('USDC', amount);
-      const sp2 = await getPricePerShare();
-      const bal2 = await getRariBalance();
-      expect(sp2.gt(sp1)).to.be.true;
-      expect(bal2.gt(bal1)).to.be.true;
-
-      // --- Withdraw ---
-      await rari.withdraw('USDC', amount);
-      const sp3 = await getPricePerShare();
-      const bal3 = await getRariBalance();
-      expect(sp3.gt(sp2)).to.be.true;
-      expect(bal3.lt(bal2)).to.be.true;
-
-      // --- Log outputs for verification ---
-      console.log('share price 1', sp1.toString());
-      console.log('share price 2', sp2.toString());
-      console.log('share price 3', sp3.toString());
-      console.log('');
-      console.log('rari receipt token balance 1', bal1.toString());
-      console.log('rari receipt token balance 2', bal2.toString());
-      console.log('rari receipt token balance 3', bal3.toString());
+      ({ deployer, ribbonVault, trigger, triggerParams } = await loadFixture(setupFixture));
     });
 
     describe('Deployment', () => {
@@ -182,7 +127,7 @@ markets.forEach((vault) => {
       it('toggles trigger when called on a broken market', async () => {
         expect(await trigger.isTriggered()).to.be.false;
 
-        await setVaultTotalSupply(MAX_UINT);
+        await setRibbonTotalSupply(MAX_UINT);
         expect(await trigger.isTriggered()).to.be.false; // trigger not updated yet, so still expect false
 
         const tx = await trigger.checkAndToggleTrigger();
@@ -197,20 +142,20 @@ markets.forEach((vault) => {
         const mockCozyToken = <MockCozyToken>await deployContract(deployer, mockCozyTokenArtifact, [trigger.address]);
         expect(await mockCozyToken.isTriggered()).to.be.false;
 
-        // Break the vault
-        await setVaultTotalSupply(MAX_UINT);
+        // Break the Ribbon vault
+        await setRibbonTotalSupply(MAX_UINT);
         await mockCozyToken.checkAndToggleTrigger();
         expect(await mockCozyToken.isTriggered()).to.be.true;
       });
 
       it('properly updates the saved state', async () => {
         // Update values
-        await setVaultTotalSupply(10n ** 18n); // don't set them too high or trigger will toggle
+        await setRibbonTotalSupply(10n ** 18n); // don't set them too high or trigger will toggle
 
         // Call checkAndToggleTrigger to simulate someone using the protocol
         await trigger.checkAndToggleTrigger();
         expect(await trigger.isTriggered()).to.be.false; // sanity check
-        const newPricePerShare = await getPricePerShare();
+        const newPricePerShare = await ribbonVault.pricePerShare();
 
         // Verify the new state
         const currentPricePerShare = await trigger.lastPricePerShare();
@@ -222,13 +167,13 @@ markets.forEach((vault) => {
         // decrease total supply by that amount
         async function modifyLastPricePerShare(numerator: bigint, denominator: bigint) {
           const lastPricePerShare = (await trigger.lastPricePerShare()).toBigInt();
-          const lastTotalSupply = (await token.totalSupply()).toBigInt();
+          const lastTotalSupply = (await ribbonVault.totalSupply()).toBigInt();
           const newTotalSupply = (lastTotalSupply * denominator) / numerator;
           const newPricePerShare = (lastPricePerShare * numerator) / denominator;
-          await setVaultTotalSupply(newTotalSupply);
+          await setRibbonTotalSupply(newTotalSupply);
           // Due to rounding error, values may sometimes differ by 1 wei, so validate with a tolerance +/2 wei
-          expect(await getPricePerShare()).to.be.above(newPricePerShare - 2n);
-          expect(await getPricePerShare()).to.be.below(newPricePerShare + 2n);
+          expect(await ribbonVault.pricePerShare()).to.be.above(newPricePerShare - 2n);
+          expect(await ribbonVault.pricePerShare()).to.be.below(newPricePerShare + 2n);
         }
 
         // Read the trigger's tolerance
